requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
pyrebase4